[MOV] bottom_bar: move bottom bar components to folder
Move the bottom bar specific components to the `components/bottom_bar` to
reduce the number of files and folders in the root of `components`.

Task: 0
