+++ src/components/bottom_bar/bottom_bar_sheet/bottom_bar_sheet.ts
@@ -1,13 +1,13 @@
 import { Component, onMounted, onPatched, useRef, useState } from "@odoo/owl";
-import { BOTTOMBAR_HEIGHT } from "../../constants";
-import { interactiveRenameSheet } from "../../helpers/ui/sheet_interactive";
-import { getSheetMenuRegistry } from "../../registries";
-import { MenuItemRegistry } from "../../registries/menu_items_registry";
-import { Store, useStore } from "../../store_engine";
-import { DOMFocusableElementStore } from "../../stores/DOM_focus_store";
-import { SpreadsheetChildEnv } from "../../types";
-import { Ripple } from "../animation/ripple";
-import { css } from "../helpers/css";
+import { BOTTOMBAR_HEIGHT } from "../../../constants";
+import { interactiveRenameSheet } from "../../../helpers/ui/sheet_interactive";
+import { getSheetMenuRegistry } from "../../../registries";
+import { MenuItemRegistry } from "../../../registries/menu_items_registry";
+import { Store, useStore } from "../../../store_engine";
+import { DOMFocusableElementStore } from "../../../stores/DOM_focus_store";
+import { SpreadsheetChildEnv } from "../../../types";
+import { Ripple } from "../../animation/ripple";
+import { css } from "../../helpers/css";
 
 css/* scss */ `
   .o-sheet {
