+++ src/components/bottom_bar/bottom_bar_statistic/aggregate_statistics_store.ts
@@ -1,9 +1,9 @@
-import { sum } from "../../functions/helper_math";
-import { average, countAny, countNumbers, max, min } from "../../functions/helper_statistical";
-import { lazy, memoize, recomputeZones } from "../../helpers";
-import { Get } from "../../store_engine";
-import { SpreadsheetStore } from "../../stores";
-import { _t } from "../../translation";
+import { sum } from "../../../functions/helper_math";
+import { average, countAny, countNumbers, max, min } from "../../../functions/helper_statistical";
+import { lazy, memoize, recomputeZones } from "../../../helpers";
+import { Get } from "../../../store_engine";
+import { SpreadsheetStore } from "../../../stores";
+import { _t } from "../../../translation";
 import {
   CellValueType,
   Command,
@@ -11,7 +11,7 @@ import {
   Lazy,
   Locale,
   invalidateEvaluationCommands,
-} from "../../types";
+} from "../../../types";
 
 export interface StatisticFnResults {
   [name: string]: Lazy<number> | undefined;
