+++ src/components/bottom_bar/bottom_bar_statistic/bottom_bar_statistic.xml
