+++ tests/bottom_bar/aggregate_statistics_store.test.ts
@@ -1,5 +1,5 @@
 import { Model } from "../../src";
-import { AggregateStatisticsStore } from "../../src/components/bottom_bar_statistic/aggregate_statistics_store";
+import { AggregateStatisticsStore } from "../../src/components/bottom_bar/bottom_bar_statistic/aggregate_statistics_store";
 import { functionRegistry } from "../../src/functions";
 import {
   activateSheet,
